+++ ImpactfulSkills/ImpactfulSkills.cs
@@ -28,7 +28,7 @@ namespace ImpactfulSkills
     {
         public const string PluginGUID = "MidnightsFX.ImpactfulSkills";
         public const string PluginName = "ImpactfulSkills";
-        public const string PluginVersion = "0.11.2";
+        public const string PluginVersion = "0.11.3";
 
         public ValConfig cfg;
         // Use this class to add your own localization to the game