+++ ImpactfulSkills/modules/Gathering.cs
@@ -208,7 +208,7 @@ namespace ImpactfulSkills.patches
                     }
 
                     Logger.LogDebug($"Checking for AOE gathering {(player_skill_factor * 100f) >= ValConfig.FarmingRangeRequiredLevel.Value} && {enabled_aoe_gathering}");
-                    if ((player_skill_factor * 100f) > ValConfig.FarmingRangeRequiredLevel.Value && enabled_aoe_gathering) {
+                    if ((player_skill_factor * 100f) >= ValConfig.FarmingRangeRequiredLevel.Value && enabled_aoe_gathering) {
                         float pickable_distance = ValConfig.GatheringRangeFactor.Value * player_skill_factor;
                         Collider[] targets = Physics.OverlapSphere(__instance.transform.position, pickable_distance, pickableMask);
                         Logger.LogDebug($"AOE Picking {targets.Count()} in harvest range {pickable_distance}.");